fix_1

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: cbukuba <cbukuba@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/21 15:39:16 by cbukuba           #+#    #+#             */
-/*   Updated: 2021/11/15 12:57:13 by cbukuba          ###   ########.fr       */
+/*   Updated: 2021/11/15 16:04:49 by cbukuba          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int			i;
+	size_t			i;
 	const char	*str1;
 	const char	*str2;
